refactor: replace code from service to controller
